requirements.txt
flwr[simulation]
torch
torchvision
PyYAML
